+++ test/integration/tests/3942-solver-error-output/files/script.hs
@@ -1,5 +1,5 @@
 #!/usr/bin/env stack
--- stack runhaskell --stack-yaml test-stack.yml --package cloud-seeder
+-- stack runhaskell --stack-yaml test-stack.yml --package one-deps
 
 main :: IO ()
 main = putStrLn "yo"